PIL implementation on Windows is now fully functional
The root of the problem (not being able to pass anything bigger than 1
byte through PIL interface) was UART overrun. When a message is big
enough, the target would overwrite RX buffer with a newer character,
while the previous one was not read yet. Setting the baud rate to 2400
fixes the problem.